refactor(ui): extract CircularIconButton and clean up TitleBar
Extract the private CircularIconButton from AppBarDefaults into a shared
top-level component with a (Dp) -> Unit content lambda. Update all call
sites in TitleBar to use the new size parameter, migrate material2 to
material3 imports in SearchInput, and adjust padding in
RegionSelectionModalContent after SearchInput lost its built-in
horizontal padding.
